Revert "hv: x2apic support for acrn"
This reverts commit 348e2ba1689f881b31309076b601a25a3339ba2e.

+++ hypervisor/include/arch/x86/per_cpu.h
@@ -47,7 +47,7 @@ struct per_cpu_region {
 	uint8_t sf_stack[CONFIG_STACK_SIZE] __aligned(16);
 	uint8_t stack[CONFIG_STACK_SIZE] __aligned(16);
 	char logbuf[LOG_MESSAGE_MAX_SIZE];
-	uint32_t lapic_id;
+	uint8_t lapic_id;
 	struct smp_call_info_data smp_call_info;
 } __aligned(CPU_PAGE_SIZE); /* per_cpu_region size aligned with CPU_PAGE_SIZE */
